Generated by Spark: This is the current driver app. We need to make the driver app more easy to follow because right now it seems like there is a lot of information that is showing up.
Another example is the today, the trips, and the rating. We need to make sure that shows today that the funds starting today and the trips starting today and the rating all in one line. We don't want it to be in two lines when it all should be in one line. It makes that section look even better if it's only one line instead of having the numbers on the top and in today at the bottom and a number on top and in trips at the bottom and in the ratings at the bottom. Don't look good, so just make it all in one line as well.