Generated by Spark: Remove the swipe left to right functionality and dig deeper to resolve this full screen request issue. It should be a pop-up. That's probably why it's not working properly.
Also, remove the swipe left to right functionality and the description that is on the side. The driver or the user of the app should not be able to scroll down or up at all on that screen. So on a full screen write request, the user should not be able to scroll up and down as well.